+++ Documents/Help/Changelog.md
@@ -6,7 +6,7 @@
 
 ## 2022-11-xx - Build 2211 - November 2022
 * `KryptonMessageBox` using the `KryptonMessageBoxIcon.Shield` will now display the correct image for the version of the OS
-* New `Apply` string for custom situations
+* New `Apply`, `Back`, `Exit`, `Finish`, `Next` & `Previous` strings for custom situations
 * Resolved [#800](https://github.com/Krypton-Suite/Standard-Toolkit/issues/800), `KryptonTextBox` not raising `KeyUp` event
 * New `Collapse` & `Expand` strings for use in expandable footers, or in other custom situations
 * Removed Visual Studio 2019 solution (*.sln) files, as 2022 is considered to be stable enough